requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
bokeh==2.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
colorcet==2.0.2
cycler==0.10.0
dask==2.14.0
datashader==0.10.0
datashape==0.5.2
decorator==4.4.2
defusedxml==0.6.0
distributed==2.14.0
entrypoints==0.3
fsspec==0.7.1
HeapDict==1.0.1
holoviews==1.13.2
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kiwisolver==1.2.0
llvmlite==0.31.0
locket==0.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
msgpack==1.0.0
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.5
networkx==2.4
notebook==6.0.3
numba==0.48.0
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
panel==0.9.5
param==1.9.3
parso==0.6.2
partd==1.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
pyct==0.4.6
Pygments==2.6.1
pynndescent==0.4.7
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pyviz-comms==0.7.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
sortedcontainers==2.1.0
tbb==2020.0.133
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typing-extensions==3.7.4.2
umap-learn==0.4.0rc2
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
xarray==0.15.1
zict==2.0.0
zipp==3.1.0
